+++ pyatv/core/mdns.py
@@ -11,8 +11,7 @@ import weakref
 from zeroconf import ServiceInfo, Zeroconf
 
 from pyatv import exceptions
-from pyatv.core import net
-from pyatv.support import log_binary
+from pyatv.support import log_binary, net
 from pyatv.support.collections import CaseInsensitiveDict
 from pyatv.support.dns import (
     DnsMessage,
