+++ pyatv/protocols/airplay/__init__.py
@@ -8,7 +8,7 @@ from typing import Any, Dict, Generator, Mapping, Optional, Set
 from pyatv import exceptions
 from pyatv.auth.hap_pairing import AuthenticationType, HapCredentials, parse_credentials
 from pyatv.const import DeviceModel, FeatureName, Protocol
-from pyatv.core import SetupData, mdns, net
+from pyatv.core import SetupData, mdns
 from pyatv.core.scan import ScanHandler, ScanHandlerReturn
 from pyatv.helpers import get_unique_id
 from pyatv.interface import (
@@ -31,6 +31,7 @@ from pyatv.protocols.airplay.pairing import (
     get_preferred_auth_type,
 )
 from pyatv.protocols.airplay.player import AirPlayPlayer
+from pyatv.support import net
 from pyatv.support.device_info import lookup_model
 from pyatv.support.http import (
     ClientSessionManager,
