+++ pyatv/protocols/mrp/connection.py
@@ -5,9 +5,9 @@ import asyncio
 import logging
 
 from pyatv import exceptions
-from pyatv.core.net import tcp_keepalive
 from pyatv.protocols.mrp import protobuf
 from pyatv.support import chacha20, log_binary, log_protobuf
+from pyatv.support.net import tcp_keepalive
 from pyatv.support.state_producer import StateProducer
 from pyatv.support.variant import read_variant, write_variant
