+++ pyatv/scripts/atvproxy.py
@@ -13,7 +13,7 @@ from zeroconf import Zeroconf
 
 from pyatv.auth.hap_srp import SRPAuthHandler
 from pyatv.const import Protocol
-from pyatv.core import mdns, net
+from pyatv.core import mdns
 from pyatv.interface import BaseService
 from pyatv.protocols.companion import opack
 from pyatv.protocols.companion.connection import CompanionConnection
@@ -23,7 +23,7 @@ from pyatv.protocols.mrp import protobuf
 from pyatv.protocols.mrp.connection import MrpConnection
 from pyatv.protocols.mrp.protocol import MrpProtocol
 from pyatv.protocols.mrp.server_auth import SERVER_IDENTIFIER, MrpServerAuth
-from pyatv.support import chacha20, log_binary, variant
+from pyatv.support import chacha20, log_binary, net, variant
 
 _LOGGER = logging.getLogger(__name__)
