+++ tests/conftest.py
@@ -8,9 +8,9 @@ import pytest
 
 import pyatv
 from pyatv.auth.hap_pairing import parse_credentials
-from pyatv.core.net import unused_port
 from pyatv.interface import BaseConfig
 from pyatv.support.http import create_session
+from pyatv.support.net import unused_port
 
 from tests import fake_udns
 from tests.fake_device.airplay import DEVICE_CREDENTIALS
