+++ tests/fake_device/dmap.py
@@ -7,8 +7,8 @@ import math
 from aiohttp import web
 
 from pyatv.const import InputAction, RepeatState, ShuffleState
-from pyatv.core.net import unused_port
 from pyatv.protocols.dmap import parser, tag_definitions, tags
+from pyatv.support.net import unused_port
 
 from tests import utils
