+++ tests/support/test_knock.py
@@ -5,8 +5,8 @@ from ipaddress import ip_address
 
 import pytest
 
-from pyatv.core.net import unused_port
 from pyatv.support.knock import knock, knocker
+from pyatv.support.net import unused_port
 
 from tests.fake_knock import create_knock_server
 from tests.utils import until
